+++ copy-dev.sh
@@ -13,7 +13,7 @@ RE="$S3_BASE$S3_BUCKET$S3_PATH"
 S3_SRC_PATH=$( echo "$@" | perl -pne 's{^.*'"$S3_BASE"'}{}; s{/index.html}{}; s{/$}{}' )
 
 S3_SRC="s3://$S3_SRC_PATH"
-S3_TARGET='s3://vitessce.io'
+S3_TARGET='s3://dev.vitessce.io'
 
 b=$(tput bold)
 n=$(tput sgr0)
@@ -25,6 +25,6 @@ then
   aws s3 cp --acl public-read \
             --recursive \
             $S3_SRC $S3_TARGET
-  open http://vitessce.io
+  open http://dev.vitessce.io
   echo 'Done. NOTE: Your browser may have cached an older verion; Reload if necessary.'
 fi