+++ src/app/view-config-utils.js
@@ -1,12 +1,13 @@
 /* eslint-disable no-plusplus */
+/* eslint-disable camelcase */
 import packageJson from '../../package.json';
 import { getNextScope } from '../utils';
 import {
-  COORDINATION_TYPES,
   DEFAULT_COORDINATION_VALUES,
   COMPONENT_COORDINATION_TYPES,
   AUTO_INDEPENDENT_COORDINATION_TYPES,
 } from './state/coordination';
+import { CoordinationType } from './constants';
 import { SCHEMA_HANDLERS } from './view-config-versions';
 
 /**
@@ -116,7 +117,7 @@ function initializeAuto(config) {
   const { layout, datasets } = newConfig;
 
   // For each coordination type, check whether it requires initialization.
-  Object.values(COORDINATION_TYPES).forEach((coordinationType) => {
+  Object.values(CoordinationType).forEach((coordinationType) => {
     // A coordination type requires coordination if at least one component is missing
     // a (coordination type, coordination scope) tuple.
     // Components may only use a subset of all coordination types.