+++ src/components/scatterplot/Scatterplot.js
@@ -118,14 +118,15 @@ class Scatterplot extends AbstractSpatialOrScatterplot {
       radiusMaxPixels: 30,
       // Our radius pixel setters measure in pixels.
       radiusUnits: 'pixels',
+      lineWidthUnits: 'pixels',
       getPosition: getCellPosition,
       getFillColor: getCellColor,
       getLineColor: getCellColor,
-      getRadius: 1,
+      getPointRadius: 1,
       getExpressionValue,
       getLineWidth: 0,
       extensions: [
-        new ScaledExpressionExtension({ instanced: true }),
+        new ScaledExpressionExtension(),
         new SelectionExtension({ instanced: true }),
       ],
       colorScaleLo: geneExpressionColormapRange[0],