+++ src/components/tooltip/Tooltip.js
@@ -1,4 +1,8 @@
-import React, { useEffect, useRef } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
+import Paper from '@material-ui/core/Paper';
+import Popper from '@material-ui/core/Popper';
+import { useVitessceContainer } from '../hooks';
+import { styles } from './styles';
 
 export default function Tooltip(props) {
   const {
@@ -8,30 +12,43 @@ export default function Tooltip(props) {
     parentHeight,
     children,
   } = props;
+  const ref = useRef();
+  const classes = styles();
+  const [placementX, setPlacementX] = useState('start');
+  const [placementY, setPlacementY] = useState('bottom');
 
-  const ref = useRef(null);
-  const isNarrow = (parentWidth < 500);
+  const getTooltipContainer = useVitessceContainer(ref);
+
   // Do collision detection based on the bounds of the tooltip ancestor element.
   useEffect(() => {
-    const el = ref.current;
-    const offsetPercentage = isNarrow ? -5 : 10;
-    const translateX = (x > parentWidth / 2) ? -(100 + offsetPercentage) : offsetPercentage;
-    const translateY = (y > parentHeight / 2) ? -(100 + offsetPercentage) : offsetPercentage;
-    const scale = isNarrow ? 0.75 : 1.0;
-    el.style.transform = `translateX(${translateX}%) translateY(${translateY}%) scale(${scale})`;
-    el.style.whiteSpace = (isNarrow ? 'normal' : 'nowrap');
-  });
+    if (ref && ref.current) {
+      const flipX = (x > parentWidth / 2);
+      const flipY = (y > parentHeight / 2);
+      setPlacementX(flipX ? 'end' : 'start');
+      setPlacementY(flipY ? 'top' : 'bottom');
+      ref.current.style.left = `${x + (flipX ? -20 : 5)}px`;
+      ref.current.style.top = `${y + (flipY ? -20 : 5)}px`;
+    }
+  }, [x, y, parentWidth, parentHeight]);
 
   return (
     <div
       ref={ref}
-      className="cell-tooltip bg-primary"
-      style={{
-        left: `${x}px`,
-        top: `${y}px`,
-      }}
+      className={classes.tooltipAnchor}
     >
-      {children}
+      {ref && ref.current ? (
+        <Popper
+          open
+          anchorEl={ref.current}
+          container={getTooltipContainer}
+          transition
+          placement={`${placementY}-${placementX}`}
+        >
+          <Paper elevation={4} className={classes.tooltipContent}>
+            {children}
+          </Paper>
+        </Popper>
+      ) : null}
     </div>
   );
 }