+++ src/layers/HeatmapCompositeTextLayer.js
@@ -1,4 +1,3 @@
-/* eslint-disable */
 /* eslint-disable no-underscore-dangle */
 import { COORDINATE_SYSTEM, CompositeLayer } from '@deck.gl/core'; // eslint-disable-line import/no-extraneous-dependencies
 import { TextLayer } from '@deck.gl/layers'; // eslint-disable-line import/no-extraneous-dependencies
@@ -8,7 +7,6 @@ import {
   AXIS_MARGIN,
   THEME_TO_TEXT_COLOR,
   AXIS_FONT_FAMILY,
-  COLOR_BAR_SIZE,
 } from './heatmap-constants';
 
 export default class HeatmapCompositeTextLayer extends CompositeLayer {
@@ -16,14 +14,14 @@ export default class HeatmapCompositeTextLayer extends CompositeLayer {
     const {
       axisTopLabelData, matrixLeft, width, matrixWidth, viewWidth, theme,
       targetX, targetY, axisTopTitle, cellWidth, axisOffsetTop, scaleFactor,
-      cellColorLabelsData, axisOffsetLeft, hideTopLabels
+      cellColorLabelsData, axisOffsetLeft, hideObservationLabels,
     } = this.props;
     const showAxisTopLabels = cellWidth >= AXIS_LABEL_TEXT_SIZE;
     const axisLabelTop = targetY + (axisOffsetTop - AXIS_MARGIN) / 2 / scaleFactor;
 
     const axisLabelLeft = targetX + (axisOffsetLeft - AXIS_MARGIN) / 2 / scaleFactor;
 
-    const topTextLayers = hideTopLabels ? [] : [
+    const topTextLayers = hideObservationLabels ? [] : [
       new TextLayer({
         id: 'axisTopLabels',
         coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
@@ -57,11 +55,11 @@ export default class HeatmapCompositeTextLayer extends CompositeLayer {
           getColor: [theme],
         },
       }),
-    ]
+    ];
 
     return [
       ...topTextLayers,
-      ...cellColorLabelsData.map((data) => (
+      ...cellColorLabelsData.map(data => (
         new TextLayer({
           id: `cellColorLabelLayer-${data[0]}`,
           coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
@@ -70,7 +68,7 @@ export default class HeatmapCompositeTextLayer extends CompositeLayer {
           getTextAnchor: 'end',
           getColor: () => THEME_TO_TEXT_COLOR[theme],
           getSize: AXIS_LABEL_TEXT_SIZE,
-          getPosition: d => [axisLabelLeft, targetY],
+          getPosition: () => [axisLabelLeft, targetY],
           getAngle: 0,
           fontFamily: AXIS_FONT_FAMILY,
         })
@@ -124,10 +122,11 @@ export default class HeatmapCompositeTextLayer extends CompositeLayer {
   }
 
   renderLayers() {
-    return [
-      this._renderAxisTopLayers(),
-      this._renderAxisLeftLayers(),
-    ];
+    const { axis } = this.props;
+    if (axis === 'left') {
+      return this._renderAxisLeftLayers();
+    }
+    return this._renderAxisTopLayers();
   }
 }